refactor Order

+++ petcalli_backend 2/src/main/java/mx/petcalli/app/controller/OrderController.java
@@ -7,14 +7,14 @@ import mx.petcalli.app.model.Order;
 public interface OrderController {
 
     
-    ResponseEntity<Order> getOrderById(Long id);
+    ResponseEntity<Order> getOrderById(Integer id);
 
     ResponseEntity<Iterable<Order> > getAllOrders ();
 
     ResponseEntity<Order> createOrder (Order newOrder);
 
-    ResponseEntity<Order> updateOrder (Order order, Long id);
+    ResponseEntity<Order> updateOrder (Order order, Integer id);
 
-    ResponseEntity<Void> deleteOrder (Long id);
+    ResponseEntity<Void> deleteOrder (Integer id);
 
 }
